+++ src/main/java/org/vaadin/addon/leaflet/shared/LeafletGridLayerState.java
@@ -1,0 +1,12 @@
+package org.vaadin.addon.leaflet.shared;
+
+public class LeafletGridLayerState extends AbstractLeafletComponentState {
+
+    public Double opacity;
+    public Integer zIndex;
+    public Bounds bounds;
+    public Integer maxNativeZoom;
+    public Integer minNativeZoom;
+    public Boolean noWrap;
+    public String attributionString;
+}
